// frontend/src/app/model/Moratoire.ts
export class Moratoire {
  morato_num: number;

	morato_sini: number;

	morato_datemiseplace: Date;

	morato_mtmoratoire: number;

	morato_typecheanc: string;

	morato_nbrecheancacc: number;

	morato_mtecheanc: number;

	morato_datech: Date;

	morato_mtnechimpayee: number;

	morato_nbrecheancimp: number;

	morato_typencdern: string;

	morato_mtnencaiss: number;

	morato_cbanq: number;

	morato_numchq: string;

	morato_dateco: Date;

	morato_cutil: string;

	morato_datemo: Date;

	morato_status: string;
}